+++ packages/geoinfo/package.js
@@ -41,6 +41,7 @@ Package.onUse(function(api) {
     'client/views/markers.html',
     'client/views/panels.html',
     'client/views/panels.js',
+    'client/views/items.html',
     'client/views/popups.html',
     'client/views/popups.js'
   ],'client');